requirements.txt
transformers==4.44.0
peft==0.12.0
trl==0.10.1
accelerate==0.34.0
bitsandbytes==0.43.3
datasets==2.21.0
torch>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0

